examples: fix the uinput example so it actually does what users expect
The previous one was technically correct, but a device with just EV_REL won't
get tagged by udev as mouse and won't get picked up by anything proper. And
likewise, without a delay before sending the event, the events won't get
picked up by userspace. Change that, so the example moves the pointer a bit
which is what a user would expect.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>